new commit

+++ ShopAPI/src/main/java/com/example/shopapi/Entity/Product.java
@@ -9,7 +9,7 @@ import java.util.List;
 public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     private String name;
     private double price;
@@ -19,11 +19,12 @@ public class Product {
     private List<Order> orders;
 
     public Product() {}
-    public int getId() {
+
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
